requirements.txt
pandas
pyyaml
matplotlib
numpy
folium
branca
h3
libpysal
esda
openpyxl